% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allmaps.R
\name{add_allmaps_layer}
\alias{add_allmaps_layer}
\title{Add an Allmaps warped map layer to a map}
\usage{
add_allmaps_layer(map, id, url, options = NULL)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl()} or \code{maplibre()} functions.}

\item{id}{A unique ID for the warped map layer.}

\item{url}{A URL or list of URLs to IIIF annotations for georeferencing, or a list of annotation objects.}

\item{options}{A list of options for the warped map layer. See details for available options.}
}
\value{
The modified map object with the warped map layer added.
}
\description{
This function adds a warped map layer using the Allmaps library to transform
historical maps onto the current map projection.
}
\details{
The options list can include:
\itemize{
\item opacity - The opacity of the warped map (0-1)
\item minzoom - Minimum zoom level at which the layer is visible
\item maxzoom - Maximum zoom level at which the layer is visible
\item bounds - An array of west, south, east, north coordinates to restrict rendering
\item attribution - Attribution text to display
}
}
\examples{
\dontrun{
library(mapgl)

# Create a map
map <- maplibre() \%>\%
  set_view(lng = 4.9, lat = 52.4, zoom = 10)

# Add a warped historical map
}
}
