% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_box_query_control}
\alias{add_box_query_control}
\title{Add a box query control to a map}
\usage{
add_box_query_control(
  map,
  position = "top-right",
  layers = NULL,
  highlighted_layer = NULL,
  filter_property = NULL,
  highlight_color = "#6e599f",
  highlight_opacity = 0.75,
  highlight_outline_color = "#484896",
  box_color = "#3887be",
  box_opacity = 0.1,
  box_border_color = "#3887be",
  box_border_width = 2,
  max_features = 1000
)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} functions.}

\item{position}{The position of the control. Can be one of "top-left", "top-right",
"bottom-left", or "bottom-right". Default is "top-right".}

\item{layers}{A character vector of layer names to include in the query. If \code{NULL} (default), all layers are queried.}

\item{highlighted_layer}{The layer ID to use for highlighting selected features. If \code{NULL}, no highlighting occurs.}

\item{filter_property}{The property name to use for filtering the highlighted layer. If \code{NULL}, the control will attempt to use feature.id or the first available property.}

\item{highlight_color}{The color to use for highlighting selected features. Default is "#6e599f".}

\item{highlight_opacity}{The opacity of highlighted features. Default is 0.75.}

\item{highlight_outline_color}{The outline color for highlighted features. Default is "#484896".}

\item{box_color}{The color of the selection box. Default is "#3887be".}

\item{box_opacity}{The opacity of the selection box. Default is 0.1.}

\item{box_border_color}{The border color of the selection box. Default is "#3887be".}

\item{box_border_width}{The border width of the selection box in pixels. Default is 2.}

\item{max_features}{Maximum number of features to select. Default is 1000.}
}
\value{
The modified map object with the box query control added.
}
\description{
This function adds a box query control to a Mapbox GL or MapLibre GL map.
The control allows users to draw a bounding box to query features within that area.
When active, box zoom is disabled and users can hold Shift+Click to draw a selection box.
}
\examples{
\dontrun{
library(mapgl)
library(sf)

# Create a map with box query control
nc <- st_read(system.file("shape/nc.shp", package = "sf"))

maplibre(style = carto_style("positron")) |>
  fit_bounds(nc) |>
  add_fill_layer(
    id = "counties",
    source = nc,
    fill_color = "lightblue",
    fill_opacity = 0.5
  ) |>
  add_fill_layer(
    id = "counties_highlighted", 
    source = nc,
    fill_color = "orange",
    fill_opacity = 0.8,
    filter = c("in", "FIPS", "")
  ) |>
  add_box_query_control(
    layers = "counties",
    highlighted_layer = "counties_highlighted",
    filter_property = "FIPS"
  )
}
}
