% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{add_categorical_legend}
\alias{add_categorical_legend}
\title{Add a categorical legend}
\usage{
add_categorical_legend(
  map,
  legend_title,
  values,
  colors,
  circular_patches = FALSE,
  position = "top-left"
)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{legend_title}{The title of the legend.}

\item{values}{The values being represented on the map (vector of categories).}

\item{colors}{The corresponding colors for the values (vector of colors).}

\item{circular_patches}{Logical, whether to use circular patches in the legend.}

\item{position}{The position of the legend on the map (one of "top-left", "bottom-left", "top-right", "bottom-right").}
}
\value{
The updated map object with the legend added.
}
\description{
Add a categorical legend
}
