% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_image_source}
\alias{add_image_source}
\title{Add an image source to a Mapbox GL map}
\usage{
add_image_source(map, id, url, coordinates)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{id}{A unique ID for the source.}

\item{url}{A URL pointing to the image source.}

\item{coordinates}{A list of coordinates specifying the image corners in clockwise order: top left, top right, bottom right, bottom left.}
}
\value{
The modified map object with the new source added.
}
\description{
Add an image source to a Mapbox GL map
}
\examples{
\dontrun{
map <- mapboxgl(access_token = "your_token_here")
map <- add_image_source(map, id = "image-source", url = "https://example.com/image.png", coordinates = list(c(-80.425, 46.437), c(-71.516, 46.437), c(-71.516, 37.936), c(-80.425, 37.936)))
}
}
