% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_source}
\alias{add_source}
\title{Add a GeoJSON or sf source to a Mapbox GL or Maplibre GL map}
\usage{
add_source(map, id, data, ...)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function.}

\item{id}{A unique ID for the source.}

\item{data}{An sf object, a \code{duckspatial_df} object, or a URL pointing to a remote GeoJSON file.}

\item{...}{Additional arguments to be passed to the JavaScript addSource method.}
}
\value{
The modified map object with the new source added.
}
\description{
Add a GeoJSON or sf source to a Mapbox GL or Maplibre GL map
}
