% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_vector_source}
\alias{add_vector_source}
\title{Add a vector tile source to a Mapbox GL or Maplibre GL map}
\usage{
add_vector_source(map, id, url = NULL, tiles = NULL, promote_id = NULL, ...)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function.}

\item{id}{A unique ID for the source.}

\item{url}{A URL pointing to the vector tile source.}

\item{tiles}{A vector of tile URLs, typically in the format "https://example.com/\{z\}/\{x\}/\{y\}.mvt" or similar.}

\item{promote_id}{An optional property name to use as the feature ID. This is required for hover effects on vector tiles.}

\item{...}{Additional arguments to be passed to the JavaScript addSource method.}
}
\value{
The modified map object with the new source added.
}
\description{
Add a vector tile source to a Mapbox GL or Maplibre GL map
}
