% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_video_source}
\alias{add_video_source}
\title{Add a video source to a Mapbox GL map}
\usage{
add_video_source(map, id, urls, coordinates)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{id}{A unique ID for the source.}

\item{urls}{A vector of URLs pointing to the video sources.}

\item{coordinates}{A list of coordinates specifying the video corners in clockwise order: top left, top right, bottom right, bottom left.}
}
\value{
The modified map object with the new source added.
}
\description{
Add a video source to a Mapbox GL map
}
\examples{
\dontrun{
map <- mapboxgl(access_token = "your_token_here")
map <- add_video_source(map, id = "video-source", urls = c("https://example.com/video.mp4", "https://example.com/video.webm"), coordinates = list(c(-80.425, 46.437), c(-71.516, 46.437), c(-71.516, 37.936), c(-80.425, 37.936)))
}
}
