% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_helpers.R
\name{carto_style}
\alias{carto_style}
\title{Get CARTO Style URL}
\usage{
carto_style(style_name)
}
\arguments{
\item{style_name}{The name of the style (e.g., "voyager", "positron", "dark-matter").}
}
\value{
The style URL corresponding to the given style name.
}
\description{
Get CARTO Style URL
}
