% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{clear_controls}
\alias{clear_controls}
\title{Clear all controls from a Mapbox GL or Maplibre GL map in a Shiny app}
\usage{
clear_controls(map, controls = NULL)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function.}
}
\value{
The modified map object with all controls removed.
}
\description{
Clear all controls from a Mapbox GL or Maplibre GL map in a Shiny app
}
