% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{clear_markers}
\alias{clear_markers}
\title{Clear markers from a map in a Shiny session}
\usage{
clear_markers(map)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl_proxy} or \code{maplibre_proxy} function.}
}
\value{
The modified map object with the markers cleared.
}
\description{
Clear markers from a map in a Shiny session
}
