% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{compareOutput}
\alias{compareOutput}
\title{Create a Compare slider widget for Shiny}
\usage{
compareOutput(outputId, type, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The output variable to read from}

\item{type}{The type of map comparison, either 'maplibregl' or 'mapboxgl'}

\item{width}{The width of the element}

\item{height}{The height of the element}
}
\value{
A Compare output element for use in a Shiny UI
}
\description{
Create a Compare slider widget for Shiny
}
