% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{ease_to}
\alias{ease_to}
\title{Ease to a given view}
\usage{
ease_to(map, center, zoom = NULL, ...)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function or a proxy object.}

\item{center}{A numeric vector of length 2 specifying the target center of the map (longitude, latitude).}

\item{zoom}{The target zoom level.}

\item{...}{Additional named arguments for easing to the view.}
}
\value{
The updated map object.
}
\description{
Ease to a given view
}
