% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hover.R
\name{enable_shiny_viewport_features}
\alias{enable_shiny_viewport_features}
\title{Enable features in viewport automatically updating shiny inputs}
\usage{
enable_shiny_viewport_features(map, layer_id)
}
\value{
The modified map object with automatic querying of viewport features
enabled.
}
\description{
This function causes a map widget in shiny to automatically refresh a shiny
input \verb{_bbox_features} with all features for the given layer in the bounding
box of the map's viewport.
}
