% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_helpers.R
\name{get_breaks}
\alias{get_breaks}
\title{Extract breaks from a classification or continuous scale object}
\usage{
get_breaks(scale)
}
\arguments{
\item{scale}{A mapgl_classification or mapgl_continuous_scale object.}
}
\value{
A numeric vector of break values.
}
\description{
This function extracts the break values from a mapgl_classification object
created by step_equal_interval(), step_quantile(), or step_jenks(), or from
a mapgl_continuous_scale object created by interpolate_palette().
}
\examples{
\dontrun{
data_values <- c(10, 25, 30, 45, 60, 75, 90)
classification <- step_jenks("value", data_values, n = 4)
breaks <- get_breaks(classification)

scale <- interpolate_palette("value", data_values, method = "equal", n = 6)
breaks <- get_breaks(scale)
}
}
