% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_helpers.R
\name{get_column}
\alias{get_column}
\title{Get column or property for use in mapping}
\usage{
get_column(column)
}
\arguments{
\item{column}{The name of the column or property to get.}
}
\value{
A list representing the expression to get the column.
}
\description{
This function returns a an expression to get a specified column from a dataset (or a property from a layer).
}
