% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_helpers.R
\name{get_legend_colors}
\alias{get_legend_colors}
\title{Extract legend colors from a classification or continuous scale object}
\usage{
get_legend_colors(scale)
}
\arguments{
\item{scale}{A mapgl_classification or mapgl_continuous_scale object.}
}
\value{
A character vector of colors.
}
\description{
This function extracts legend colors from a mapgl_classification object
created by step_equal_interval(), step_quantile(), or step_jenks(), or from
a mapgl_continuous_scale object created by interpolate_palette().
}
\examples{
\dontrun{
data_values <- c(10, 25, 30, 45, 60, 75, 90)
classification <- step_equal_interval("value", data_values, n = 4)
colors <- get_legend_colors(classification)

scale <- interpolate_palette("value", data_values, method = "jenks", n = 5)
colors <- get_legend_colors(scale)
}
}
