% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{jump_to}
\alias{jump_to}
\title{Jump to a given view}
\usage{
jump_to(map, center, zoom = NULL, ...)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function or a proxy object.}

\item{center}{A numeric vector of length 2 specifying the target center of the map (longitude, latitude).}

\item{zoom}{The target zoom level.}

\item{...}{Additional named arguments for jumping to the view.}
}
\value{
The updated map object.
}
\description{
Jump to a given view
}
