% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_helpers.R
\name{mapbox_style}
\alias{mapbox_style}
\title{Get Mapbox Style URL}
\usage{
mapbox_style(style_name)
}
\arguments{
\item{style_name}{The name of the style (e.g., "standard", "streets", "outdoors", etc.).}
}
\value{
The style URL corresponding to the given style name.
}
\description{
Get Mapbox Style URL
}
