% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxgl-package.R
\docType{package}
\name{mapboxgl-package}
\alias{mapboxgl-package}
\title{mapboxgl: Wrapper for Mapbox GL JS v3}
\description{
Provides an interface to Mapbox GL JS v3 for interactive, customizable maps in R.
}
\author{
\strong{Maintainer}: Kyle Walker \email{kyle@walker-data.com}

}
\keyword{internal}
