% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{mapboxgl_proxy}
\alias{mapboxgl_proxy}
\title{Create a proxy object for a Mapbox GL map in Shiny}
\usage{
mapboxgl_proxy(mapId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{mapId}{The ID of the map output element.}

\item{session}{The Shiny session object.}
}
\value{
A proxy object for the Mapbox GL map.
}
\description{
This function allows updates to be sent to an existing Mapbox GL map in a Shiny application without redrawing the entire map.
}
