% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxgl.R
\name{mapboxglOutput}
\alias{mapboxglOutput}
\title{Create a Mapbox GL output element for Shiny}
\usage{
mapboxglOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The output variable to read from}

\item{width}{The width of the element}

\item{height}{The height of the element}
}
\value{
A Mapbox GL output element for use in a Shiny UI
}
\description{
Create a Mapbox GL output element for Shiny
}
