% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{maplibre_proxy}
\alias{maplibre_proxy}
\title{Create a proxy object for a Maplibre GL map in Shiny}
\usage{
maplibre_proxy(mapId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{mapId}{The ID of the map output element.}

\item{session}{The Shiny session object.}
}
\value{
A proxy object for the Maplibre GL map.
}
\description{
This function allows updates to be sent to an existing Maplibre GL map in a Shiny application without redrawing the entire map.
}
