% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{maplibreCompareOutput}
\alias{maplibreCompareOutput}
\title{Create a Maplibre GL Compare output element for Shiny}
\usage{
maplibreCompareOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The output variable to read from}

\item{width}{The width of the element}

\item{height}{The height of the element}
}
\value{
A Maplibre GL Compare output element for use in a Shiny UI
}
\description{
Create a Maplibre GL Compare output element for Shiny
}
