% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplibre.R
\name{maplibreOutput}
\alias{maplibreOutput}
\title{Create a Maplibre GL output element for Shiny}
\usage{
maplibreOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The output variable to read from}

\item{width}{The width of the element}

\item{height}{The height of the element}
}
\value{
A Maplibre GL output element for use in a Shiny UI
}
\description{
Create a Maplibre GL output element for Shiny
}
