% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_helpers.R
\name{maptiler_style}
\alias{maptiler_style}
\title{Get MapTiler Style URL}
\usage{
maptiler_style(style_name, api_key = NULL)
}
\arguments{
\item{style_name}{The name of the style (e.g., "basic", "streets", "toner", etc.).}

\item{api_key}{Your MapTiler API key (required)}
}
\value{
The style URL corresponding to the given style name.
}
\description{
Get MapTiler Style URL
}
