% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_helpers.R
\name{openfreemap_style}
\alias{openfreemap_style}
\title{Get OpenFreeMap Style URL}
\usage{
openfreemap_style(style_name)
}
\arguments{
\item{style_name}{The name of the style (e.g., "bright", "positron", "liberty").}
}
\value{
The style URL corresponding to the given style name.
}
\description{
Get OpenFreeMap Style URL
}
