% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{renderCompare}
\alias{renderCompare}
\title{Render a Compare output element in Shiny}
\usage{
renderCompare(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a Compare element}

\item{env}{The environment in which to evaluate \code{expr}}

\item{quoted}{Is \code{expr} a quoted expression}
}
\value{
A rendered Compare element for use in a Shiny server
}
\description{
Render a Compare output element in Shiny
}
