% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxgl.R
\name{renderMapboxgl}
\alias{renderMapboxgl}
\title{Render a Mapbox GL output element in Shiny}
\usage{
renderMapboxgl(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a Mapbox GL map}

\item{env}{The environment in which to evaluate \code{expr}}

\item{quoted}{Is \code{expr} a quoted expression}
}
\value{
A rendered Mapbox GL map for use in a Shiny server
}
\description{
Render a Mapbox GL output element in Shiny
}
