% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{renderMapboxglCompare}
\alias{renderMapboxglCompare}
\title{Render a Mapbox GL Compare output element in Shiny}
\usage{
renderMapboxglCompare(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a Mapbox GL Compare map}

\item{env}{The environment in which to evaluate \code{expr}}

\item{quoted}{Is \code{expr} a quoted expression}
}
\value{
A rendered Mapbox GL Compare map for use in a Shiny server
}
\description{
Render a Mapbox GL Compare output element in Shiny
}
