% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplibre.R
\name{renderMaplibre}
\alias{renderMaplibre}
\title{Render a Maplibre GL output element in Shiny}
\usage{
renderMaplibre(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a Maplibre GL map}

\item{env}{The environment in which to evaluate \code{expr}}

\item{quoted}{Is \code{expr} a quoted expression}
}
\value{
A rendered Maplibre GL map for use in a Shiny server
}
\description{
Render a Maplibre GL output element in Shiny
}
