% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{renderMaplibreCompare}
\alias{renderMaplibreCompare}
\title{Render a Maplibre GL Compare output element in Shiny}
\usage{
renderMaplibreCompare(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a Maplibre GL Compare map}

\item{env}{The environment in which to evaluate \code{expr}}

\item{quoted}{Is \code{expr} a quoted expression}
}
\value{
A rendered Maplibre GL Compare map for use in a Shiny server
}
\description{
Render a Maplibre GL Compare output element in Shiny
}
