% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_helpers.R
\name{set_config_property}
\alias{set_config_property}
\title{Set a configuration property for the Mapbox GL map}
\usage{
set_config_property(map, import_id, config_name, value)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function or a proxy object defined with \code{mapboxgl_proxy()}.}

\item{import_id}{The name of the imported style to set the config for (e.g., 'basemap').}

\item{config_name}{The name of the configuration property from the style.}

\item{value}{The value to set for the configuration property.}
}
\value{
The updated map object with the configuration property set.
}
\description{
Set a configuration property for the Mapbox GL map
}
