% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{set_paint_property}
\alias{set_paint_property}
\title{Set a paint property on a map layer}
\usage{
set_paint_property(map, layer_id = NULL, name, value, layer = NULL)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function, or a proxy object.}

\item{layer_id}{The ID of the layer to update.}

\item{name}{The name of the paint property to set.}

\item{value}{The value to set the property to.}

\item{layer}{Deprecated. Use \code{layer_id} instead.}
}
\value{
The updated map object.
}
\description{
Set a paint property on a map layer
}
