% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_helpers.R
\name{set_projection}
\alias{set_projection}
\title{Set Projection for a Mapbox/Maplibre Map}
\usage{
set_projection(map, projection)
}
\arguments{
\item{map}{A map object created by mapboxgl() or maplibre() functions, or their respective proxy objects}

\item{projection}{A string representing the projection name (e.g., "mercator", "globe", "albers", "equalEarth", etc.)}
}
\value{
The modified map object
}
\description{
This function sets the projection dynamically after map initialization.
}
