% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{set_view}
\alias{set_view}
\title{Set the map center and zoom level}
\usage{
set_view(map, center, zoom)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function or a proxy object.}

\item{center}{A numeric vector of length 2 specifying the center of the map (longitude, latitude).}

\item{zoom}{The zoom level.}
}
\value{
The updated map object.
}
\description{
Set the map center and zoom level
}
