% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_helpers.R
\name{step_jenks}
\alias{step_jenks}
\title{Create a step expression with Jenks natural breaks classification}
\usage{
step_jenks(column, data_values, n = 5, colors = NULL, na_color = "grey")
}
\arguments{
\item{column}{The name of the column to use for the step expression.}

\item{data_values}{A numeric vector of the actual data values used to calculate breaks.}

\item{n}{The number of classes to create. Defaults to 5.}

\item{colors}{A vector of colors to use. If NULL, uses viridisLite::viridis(n).}

\item{na_color}{The color to use for missing values. Defaults to "grey".}
}
\value{
A list of class "mapgl_classification" containing the step expression and metadata.
}
\description{
This function creates a step expression using Jenks natural breaks (also known as
Fisher-Jenks optimal classification). This method minimizes within-class variance
while maximizing between-class variance, similar to the natural breaks option
in GIS software like ArcGIS.
}
\examples{
\dontrun{
# Create Jenks natural breaks classification
data_values <- c(2, 5, 8, 12, 20, 25, 40, 60, 80, 100)
classification <- step_jenks("value", data_values, n = 4)

# Use in a layer
add_circle_layer(map, circle_color = classification$expression)

# Extract legend information
labels <- classification$labels
colors <- classification$colors
}
}
