% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storymaps.R
\name{story_map}
\alias{story_map}
\title{Create a scrollytelling story map}
\usage{
story_map(map_id, sections, root_margin = "-20\% 0px -20\% 0px", styles = NULL)
}
\arguments{
\item{map_id}{The ID of your mapboxgl/maplibre output}

\item{sections}{Named list of story_section objects}

\item{root_margin}{Margin around viewport for triggering sections}

\item{styles}{Optional custom CSS styles}
}
\description{
Create a scrollytelling story map
}
