% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{turf_distance}
\alias{turf_distance}
\title{Calculate distance between two features}
\usage{
turf_distance(
  proxy,
  layer_id,
  layer_id_2,
  units = "kilometers",
  send_to_r = TRUE
)
}
\arguments{
\item{proxy}{A mapboxgl_proxy or maplibre_proxy object.}

\item{layer_id}{The ID of the first layer.}

\item{layer_id_2}{The ID of the second layer.}

\item{units}{The units for the distance calculation. One of "meters", "kilometers", "miles", etc.}

\item{send_to_r}{Logical. If TRUE, sends the result back to R as input$<map_id>_turf_result. Default is TRUE.}
}
\value{
The proxy object for method chaining.
}
\description{
This function calculates the distance between the first features of two layers.
}
