% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{update_feature_query}
\alias{update_feature_query}
\title{Get a set of features on a map}
\usage{
update_feature_query(map, layer_id = NULL, geometry = NULL)
}
\arguments{
\item{layer_id}{layer_id of map layer from which to query features.}

\item{geometry}{if NULL, uses the bbox of the current viewport. Otherwise, a
specific geometry can be specified to pull features overlapping that
geometry.

See \link[=https://maplibre.org/maplibre-gl-js/docs/API/classes/Map/#queryrenderedfeatures]{queryrenderedfeatures}}
}
\value{
Nothing, but updates \code{input$map_feature_query}
}
\description{
Get a set of features on a map
}
