% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{update_source_query}
\alias{update_source_query}
\title{Queries  all features in a source}
\usage{
update_source_query(map, source_id, source_layer = NULL)
}
\arguments{
\item{source_id}{id of source layer for which to query features. Note the
returned features will be empty if this is left NULL and query is refering
to a basemap layer.

See
\link[=https://maplibre.org/maplibre-gl-js/docs/API/classes/Map/#querysourcefeatures]{querySourceFeatures}}

\item{layer_id}{layer_id of map layer from which to query features.}
}
\value{
Nothing, but updates \code{input$map_source_query} in shiny contexts.
}
\description{
Will pull all features (visible or not) from a given source and save as a
shiny input
}
