% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_circle_layer}
\alias{add_circle_layer}
\title{Add a circle layer to a Mapbox GL map}
\usage{
add_circle_layer(
  map,
  id,
  source,
  source_layer = NULL,
  circle_blur = NULL,
  circle_color = NULL,
  circle_opacity = NULL,
  circle_radius = NULL,
  circle_sort_key = NULL,
  circle_stroke_color = NULL,
  circle_stroke_opacity = NULL,
  circle_stroke_width = NULL,
  circle_translate = NULL,
  circle_translate_anchor = "map",
  visibility = "visible",
  slot = NULL,
  min_zoom = NULL,
  max_zoom = NULL,
  popup = NULL,
  tooltip = NULL
)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{id}{A unique ID for the layer.}

\item{source}{The ID of the source, alternatively an sf object (which will be converted to a GeoJSON source) or a named list that specifies \code{type} and \code{url} for a remote source.}

\item{source_layer}{The source layer (for vector sources).}

\item{circle_blur}{Amount to blur the circle.}

\item{circle_color}{The color of the circle.}

\item{circle_opacity}{The opacity at which the circle will be drawn.}

\item{circle_radius}{Circle radius.}

\item{circle_sort_key}{Sorts features in ascending order based on this value.}

\item{circle_stroke_color}{The color of the circle's stroke.}

\item{circle_stroke_opacity}{The opacity of the circle's stroke.}

\item{circle_stroke_width}{The width of the circle's stroke.}

\item{circle_translate}{The geometry's offset. Values are \link{x, y} where negatives indicate left and up.}

\item{circle_translate_anchor}{Controls the frame of reference for \code{circle-translate}.}

\item{visibility}{Whether this layer is displayed.}

\item{slot}{An optional slot for layer order.}

\item{min_zoom}{The minimum zoom level for the layer.}

\item{max_zoom}{The maximum zoom level for the layer.}

\item{popup}{A column name containing information to display in a popup on click.  Columns containing HTML will be parsed.}

\item{tooltip}{A column name containing information to display in a tooltip on hover. Columns containing HTML will be parsed.}
}
\value{
The modified map object with the new circle layer added.
}
\description{
Add a circle layer to a Mapbox GL map
}
\examples{
\dontrun{
map <- mapboxgl(
  style = "mapbox://styles/mapbox/streets-v11",
  center = c(-74.006, 40.7128),
  zoom = 10,
  access_token = "your_token_here"
)
map <- add_circle_layer(map, id = "circle-layer", source = "source-id", circle_color = "rgba(255,0,0,0.5)")
}
}
