% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{add_continuous_legend}
\alias{add_continuous_legend}
\title{Add a continuous legend}
\usage{
add_continuous_legend(map, legend_title, values, colors)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{legend_title}{The title of the legend.}

\item{values}{The values being represented on the map (vector of stops).}

\item{colors}{The colors used to generate the color ramp.}
}
\value{
The updated map object with the legend added.
}
\description{
Add a continuous legend
}
