% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_draw_control}
\alias{add_draw_control}
\title{Add a draw control to a map}
\usage{
add_draw_control(map, position = "top-left")
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} functions.}

\item{position}{A string specifying the position of the draw control.
One of "top-right", "top-left", "bottom-right", or "bottom-left".}
}
\value{
The modified map object with the draw control added.
}
\description{
Add a draw control to a map
}
\examples{
\dontrun{
library(mapgl)

mapboxgl(style = mapbox_style("streets"),
         center = c(-74.50, 40),
         zoom = 9) |>
  add_draw_control(position = "top-left")
}
}
