% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_fullscreen_control}
\alias{add_fullscreen_control}
\title{Add a fullscreen control to a Mapbox GL map}
\usage{
add_fullscreen_control(map)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}
}
\value{
The modified map object with the fullscreen control added.
}
\description{
Add a fullscreen control to a Mapbox GL map
}
\examples{
\dontrun{
map <- mapboxgl(
  style = "mapbox://styles/mapbox/outdoors-v12",
  center = c(11.255, 43.77),
  zoom = 13,
  access_token = "your_token_here"
)
map <- add_fullscreen_control(map)
}
}
