% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeferenced_map.R
\name{add_georeferenced_map}
\alias{add_georeferenced_map}
\title{Add an Allmaps georeferenced map layer}
\usage{
add_georeferenced_map(
  map,
  url,
  id = NULL,
  opacity = NULL,
  colorize = NULL,
  remove_color = list(hex_color = NULL, threshold = NULL, hardness = NULL),
  saturation = NULL,
  before_id = NULL
)
}
\arguments{
\item{map}{A map object created by the \code{maplibre()} functions.}

\item{url}{An Allmaps URL or list of URLs. (Required).}

\item{id}{An optional unique ID for the warped map layer. If NULL, it will
default to "warped-map-layer".}

\item{opacity}{An optional numeric between 0 and 1. Controls the opacity of
the warped map layer. If NULL, it will default to 1.}

\item{colorize}{An optional colorization to apply to the map. Must be a hex code.}

\item{remove_color}{An optional list containing options for color removal:
\itemize{
\item \code{hex_color}: The color to remove (hex code)
\item \code{threshold}: Numeric threshold for color removal (0-1)
\item \code{hardness}: Hardness of the color removal effect (0-1)
}}

\item{saturation}{An optional numeric between 0 and 1. Adjusts the saturation
of the warped map, where 0 is grayscale and 1 is full color. If NULL, it will
default to 1.}

\item{before_id}{The name of the layer that this layer appears "before",
allowing you to insert layers below other layers in your basemap (e.g. labels).}
}
\value{
The modified map object with the warped map layer added.
}
\description{
This function adds a georeferenced map from \href{https://allmaps.org/}{Allmaps.org}
as a layer to the current map. For further information and more-thorough
documentation of the plugin, visit the
\href{https://github.com/allmaps/allmaps/tree/develop/packages/maplibre}{@allmaps/maplibre website}
}
\examples{
\dontrun{
library(mapgl)

# Add a georeferenced map with a unique ID, 75\% opacity,
#   and colorized to blue
maplibre(center = c(-73.95, 40.75), zoom = 10) |>
  add_georeferenced_map(
    id = "old-manhattan",
    url = "https://annotations.allmaps.org/images/d180902cb93d5bf2",
    opacity = 0.75,
    colorize = "#0000FF"
  )
# Add a georeferenced map with background gray removed at custom threshold
#  and hardness levels, then converted to grayscale
maplibre(center = c(-73.95, 40.75), zoom = 10) |>
  add_georeferenced_map(
    url = "https://annotations.allmaps.org/images/d180902cb93d5bf2",
    remove_color = list(
      hex_color = "#b7c3b8",
      threshold = 0.3,
      hardness = 0.5
    ),
    saturation = 0
  )
}
}
\seealso{
\itemize{
\item \code{\link[=maplibre]{maplibre()}} for creating the base map
\item \code{\link[=add_layer]{add_layer()}} for adding other types of layers
}
}
