% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_heatmap_layer}
\alias{add_heatmap_layer}
\title{Add a heatmap layer to a Mapbox GL map}
\usage{
add_heatmap_layer(
  map,
  id,
  source,
  source_layer = NULL,
  heatmap_color = NULL,
  heatmap_intensity = NULL,
  heatmap_opacity = NULL,
  heatmap_radius = NULL,
  heatmap_weight = NULL,
  visibility = "visible",
  slot = NULL,
  min_zoom = NULL,
  max_zoom = NULL
)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{id}{A unique ID for the layer.}

\item{source}{The ID of the source, alternatively an sf object (which will be converted to a GeoJSON source) or a named list that specifies \code{type} and \code{url} for a remote source.}

\item{source_layer}{The source layer (for vector sources).}

\item{heatmap_color}{The color of the heatmap points.}

\item{heatmap_intensity}{The intensity of the heatmap points.}

\item{heatmap_opacity}{The opacity of the heatmap layer.}

\item{heatmap_radius}{The radius of influence of each individual heatmap point.}

\item{heatmap_weight}{The weight of each individual heatmap point.}

\item{visibility}{Whether this layer is displayed.}

\item{slot}{An optional slot for layer order.}

\item{min_zoom}{The minimum zoom level for the layer.}

\item{max_zoom}{The maximum zoom level for the layer.}
}
\value{
The modified map object with the new heatmap layer added.
}
\description{
Add a heatmap layer to a Mapbox GL map
}
\examples{
\dontrun{
map <- mapboxgl(
  style = "mapbox://styles/mapbox/streets-v11",
  center = c(-74.006, 40.7128),
  zoom = 10,
  access_token = "your_token_here"
)
map <- add_heatmap_layer(map, id = "heatmap-layer", source = "source-id", heatmap_color = "rgba(255,0,0,0.5)")
}
}
