% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_image_source}
\alias{add_image_source}
\title{Add an image source to a Mapbox GL or Maplibre GL map}
\usage{
add_image_source(map, id, url, coordinates)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function.}

\item{id}{A unique ID for the source.}

\item{url}{A URL pointing to the image source.}

\item{coordinates}{A list of coordinates specifying the image corners in clockwise order: top left, top right, bottom right, bottom left.}
}
\value{
The modified map object with the new source added.
}
\description{
Add an image source to a Mapbox GL or Maplibre GL map
}
