% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_layer}
\alias{add_layer}
\title{Add a layer to a Mapbox GL map from a source}
\usage{
add_layer(
  map,
  id,
  type = "fill",
  source,
  source_layer = NULL,
  paint = list(),
  layout = list(),
  slot = NULL,
  min_zoom = NULL,
  max_zoom = NULL,
  popup = NULL,
  tooltip = NULL,
  hover_options = NULL
)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{id}{A unique ID for the layer.}

\item{type}{The type of the layer (e.g., "fill", "line", "circle").}

\item{source}{The ID of the source, alternatively an sf object (which will be converted to a GeoJSON source) or a named list that specifies \code{type} and \code{url} for a remote source.}

\item{source_layer}{The source layer (for vector sources).}

\item{paint}{A list of paint properties for the layer.}

\item{layout}{A list of layout properties for the layer.}

\item{slot}{An optional slot for layer order.}

\item{min_zoom}{The minimum zoom level for the layer.}

\item{max_zoom}{The maximum zoom level for the layer.}

\item{popup}{A column name containing information to display in a popup on click.  Columns containing HTML will be parsed.}

\item{tooltip}{A column name containing information to display in a tooltip on hover. Columns containing HTML will be parsed.}

\item{hover_options}{A named list of options for highlighting features in the layer on hover.}
}
\value{
The modified map object with the new layer added.
}
\description{
Add a layer to a Mapbox GL map from a source
}
\examples{
\dontrun{
map <- mapboxgl(access_token = "your_token_here")
map <- add_source(map, id = "nc-source", type = "geojson", data = nc)
map <- add_layer(map, id = "nc-layer", type = "fill", source = "nc-source", paint = list("fill-color" = "#888888", "fill-opacity" = 0.4))
map <- add_source(map, id = "external-source", type = "vector", url = "mapbox://mapbox.mapbox-streets-v8")
map <- add_layer(map, id = "external-layer", type = "circle", source = "external-source", source_layer = "sf2010", paint = list("circle-radius" = list("base" = 1.75, "stops" = list(c(12, 2), c(22, 180))), "circle-color" = list("match", list("get", "ethnicity"), "White", "#fbb03b", "Black", "#223b53", "Hispanic", "#e55e5e", "Asian", "#3bb2d0", "other", "#ccc")))
}
}
