% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{add_legend}
\alias{add_legend}
\title{Add a legend to a Mapbox GL map}
\usage{
add_legend(
  map,
  legend_title,
  values,
  colors,
  type = c("continuous", "categorical"),
  circular_patches = FALSE
)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{legend_title}{The title of the legend.}

\item{values}{The values being represented on the map (either a vector of categories or a vector of stops).}

\item{colors}{The corresponding colors for the values (either a vector of colors or an interpolate function).}

\item{type}{one of "continuous" or "categorical"}

\item{circular_patches}{Logical, whether to use circular patches in the legend.}
}
\value{
The updated map object with the legend added.
}
\description{
Add a legend to a Mapbox GL map
}
