% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{add_legend}
\alias{add_legend}
\title{Add a legend to a Mapbox GL map}
\usage{
add_legend(
  map,
  legend_title,
  values,
  colors,
  type = c("continuous", "categorical"),
  circular_patches = FALSE,
  position = "top-left",
  sizes = NULL,
  add = FALSE
)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{legend_title}{The title of the legend.}

\item{values}{The values being represented on the map (either a vector of categories or a vector of stops).}

\item{colors}{The corresponding colors for the values (either a vector of colors, a single color, or an interpolate function).}

\item{type}{One of "continuous" or "categorical".}

\item{circular_patches}{Logical, whether to use circular patches in the legend (only for categorical legends).}

\item{position}{The position of the legend on the map (one of "top-left", "bottom-left", "top-right", "bottom-right").}

\item{sizes}{An optional numeric vector of sizes for the legend patches, or a single numeric value (only for categorical legends).}

\item{add}{Logical, whether to add this legend to existing legends (TRUE) or replace existing legends (FALSE). Default is FALSE.}
}
\value{
The updated map object with the legend added.
}
\description{
Add a legend to a Mapbox GL map
}
