% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{add_markers}
\alias{add_markers}
\title{Add markers to a Mapbox GL map}
\usage{
add_markers(map, data, color = "red", rotation = 0, popup = NULL, ...)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{data}{A length-2 numeric vector of coordinates or an \code{sf} POINT object.}

\item{color}{The color of the marker.}

\item{rotation}{The rotation of the marker.}

\item{popup}{A column name for popups (if data is an \code{sf} object) or a string for a single popup (if data is a numeric vector).}

\item{...}{Additional options passed to the marker.}
}
\value{
The modified map object with the markers added.
}
\description{
Add markers to a Mapbox GL map
}
\examples{
\dontrun{
map <- mapboxgl(
  style = "mapbox://styles/mapbox/streets-v12",
  center = c(-74.006, 40.7128),
  zoom = 10,
  access_token = "your_token_here"
)

# Add a single marker
map <- add_markers(map, c(-74.006, 40.7128), color = "blue", rotation = 45, popup = "A marker")

# Add multiple markers from an sf object
map <- add_markers(map, your_sf_object, color = "red", popup = "your_column_name")
}
}
