% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_pmtiles_source}
\alias{add_pmtiles_source}
\title{Add a PMTiles source to a Mapbox GL or Maplibre GL map}
\usage{
add_pmtiles_source(map, id, url, ...)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function.}

\item{id}{A unique ID for the source.}

\item{url}{A URL pointing to the PMTiles archive.}

\item{...}{Additional arguments to be passed to the JavaScript addSource method.}
}
\value{
The modified map object with the new source added.
}
\description{
Add a PMTiles source to a Mapbox GL or Maplibre GL map
}
\examples{
\dontrun{
# For MapLibre GL JS (uses native PMTiles support)
maplibre()
  add_pmtiles_source(
    id = "pmtiles-source",
    url = "https://example.com/data.pmtiles"
  ) |>
  add_layer(
    id = "pmtiles-layer",
    source = "pmtiles-source",
    source_layer = "layer-name",
    type = "fill",
    paint = list(
      "fill-color" = "blue"
    )
  )

# For Mapbox GL JS (uses custom PMTiles source type)
mapboxgl() |>
  add_pmtiles_source(
    id = "pmtiles-source",
    url = "https://example.com/data.pmtiles"
  ) |>
  add_layer(
    id = "pmtiles-layer",
    source = "pmtiles-source",
    source_layer = "layer-name",
    type = "fill",
    paint = list(
      "fill-color" = "blue"
    )
  )
}
}
