% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_source}
\alias{add_source}
\title{Add a GeoJSON source to a Mapbox GL map}
\usage{
add_source(map, id, data)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} function.}

\item{id}{A unique ID for the source.}

\item{data}{An sf object or a URL pointing to a remote GeoJSON file.}
}
\value{
The modified map object with the new source added.
}
\description{
Add a GeoJSON source to a Mapbox GL map
}
\examples{
\dontrun{
nc <- st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
map <- mapboxgl(access_token = "your_token_here")
map <- add_source(map, id = "nc-source", data = nc)
}
}
