% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_video_source}
\alias{add_video_source}
\title{Add a video source to a Mapbox GL or Maplibre GL map}
\usage{
add_video_source(map, id, urls, coordinates)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function.}

\item{id}{A unique ID for the source.}

\item{urls}{A vector of URLs pointing to the video sources.}

\item{coordinates}{A list of coordinates specifying the video corners in clockwise order: top left, top right, bottom right, bottom left.}
}
\value{
The modified map object with the new source added.
}
\description{
Add a video source to a Mapbox GL or Maplibre GL map
}
