% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{clear_layer}
\alias{clear_layer}
\title{Clear a layer from a map using a proxy}
\usage{
clear_layer(proxy, layer_id)
}
\arguments{
\item{proxy}{A proxy object created by \code{mapboxgl_proxy} or \code{maplibre_proxy}.}

\item{layer_id}{The ID of the layer to be removed.}
}
\value{
The updated proxy object.
}
\description{
This function allows a layer to be removed from an existing Mapbox GL map using a proxy object.
}
