% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{clear_legend}
\alias{clear_legend}
\title{Clear legend from a map in a proxy session}
\usage{
clear_legend(map)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl_proxy} or \code{maplibre_proxy} function.}
}
\value{
The updated map object with the legend cleared.
}
\description{
Clear legend from a map in a proxy session
}
