% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{compareOutput}
\alias{compareOutput}
\title{Create a UI output for a comparison map}
\usage{
compareOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The output variable to read the map from.}

\item{width}{The width of the map container.}

\item{height}{The height of the map container.}
}
\value{
A UI output element for the comparison map.
}
\description{
This function creates a UI output element for a comparison map in a Shiny application.
}
