% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{fit_bounds}
\alias{fit_bounds}
\title{Fit the map to a bounding box}
\usage{
fit_bounds(map, bbox, ...)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function or a proxy object.}

\item{bbox}{A bounding box specified as a numeric vector of length 4 (minLng, minLat, maxLng, maxLat), or an sf object from which a bounding box will be calculated.}

\item{...}{Additional named arguments for fitting the bounds.}
}
\value{
The updated map object.
}
\description{
Fit the map to a bounding box
}
