% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{mapboxgl_compare_proxy}
\alias{mapboxgl_compare_proxy}
\title{Create a proxy object for a Mapbox GL Compare widget in Shiny}
\usage{
mapboxgl_compare_proxy(
  compareId,
  session = shiny::getDefaultReactiveDomain(),
  map_side = "before"
)
}
\arguments{
\item{compareId}{The ID of the compare output element.}

\item{session}{The Shiny session object.}

\item{map_side}{Which map side to target in the compare widget, either "before" or "after".}
}
\value{
A proxy object for the Mapbox GL Compare widget.
}
\description{
This function allows updates to be sent to an existing Mapbox GL Compare widget in a Shiny application.
}
