% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapgl-package.R
\docType{package}
\name{mapgl-package}
\alias{mapgl}
\alias{mapgl-package}
\title{mapgl: Interactive Maps With 'Mapbox GL JS' and 'MapLibre GL JS' in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an interface to the 'Mapbox GL JS' (\url{https://docs.mapbox.com/mapbox-gl-js/guides}) and the 'MapLibre GL JS' (\url{https://maplibre.org/maplibre-gl-js/docs/}) interactive mapping libraries to help users create custom interactive maps in R. Users can create interactive globe visualizations; layer 'sf' objects to create filled maps, circle maps, 'heatmaps', and three-dimensional graphics; and customize map styles and views. The package also includes utilities to use 'Mapbox' and 'MapLibre' maps in 'Shiny' web applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://walker-data.com/mapgl/}
}

}
\author{
\strong{Maintainer}: Kyle Walker \email{kyle@walker-data.com}

}
\keyword{internal}
