% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplibre.R
\name{maplibre}
\alias{maplibre}
\title{Initialize a Maplibre GL Map}
\usage{
maplibre(
  style = carto_style("voyager"),
  center = c(0, 0),
  zoom = 0,
  bearing = 0,
  pitch = 0,
  bounds = NULL,
  width = "100\%",
  height = NULL,
  ...
)
}
\arguments{
\item{style}{The style JSON to use.}

\item{center}{A numeric vector of length 2 specifying the initial center of the map.}

\item{zoom}{The initial zoom level of the map.}

\item{bearing}{The initial bearing (rotation) of the map, in degrees.}

\item{pitch}{The initial pitch (tilt) of the map, in degrees.}

\item{bounds}{An sf object or an unnamed numeric vector of the form \code{c(xmin, ymin, xmax, ymax)}
indicating the bounding box to fit the map to.}

\item{width}{The width of the output htmlwidget.}

\item{height}{The height of the output htmlwidget.}

\item{...}{Additional named parameters to be passed to the Mapbox GL map.}
}
\value{
An HTML widget for a Mapbox map.
}
\description{
Initialize a Maplibre GL Map
}
\examples{
\dontrun{
maplibre()
}
}
