% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{move_layer}
\alias{move_layer}
\title{Move a layer to a different z-position}
\usage{
move_layer(proxy, layer_id, before_id = NULL)
}
\arguments{
\item{proxy}{A proxy object created by \code{mapboxgl_proxy} or \code{maplibre_proxy}.}

\item{layer_id}{The ID of the layer to move.}

\item{before_id}{The ID of an existing layer to insert the new layer before. \strong{Important}: this means that the layer will appear \emph{immediately behind} the layer defined in \code{before_id}. If omitted, the layer will be appended to the end of the layers array and appear above all other layers.}
}
\value{
The updated proxy object.
}
\description{
This function allows a layer to be moved to a different z-position in an existing Mapbox GL or Maplibre GL map using a proxy object.
}
