% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storymaps.R
\name{on_section}
\alias{on_section}
\title{Handle story map section visibility}
\usage{
on_section(map_id, section_id, handler)
}
\arguments{
\item{map_id}{The ID of your mapboxgl/maplibre output}

\item{section_id}{The ID of the section to trigger on}

\item{handler}{Expression to execute when section becomes visible}
}
\description{
Handle story map section visibility
}
