% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{query_rendered_features}
\alias{query_rendered_features}
\title{Query rendered features on a map in a Shiny session}
\usage{
query_rendered_features(proxy, geometry = NULL, layers = NULL, filter = NULL)
}
\arguments{
\item{proxy}{A MapboxGL or Maplibre proxy object, defined with \code{mapboxgl_proxy()} or \code{maplibre_proxy()}}

\item{geometry}{The geometry to query. Should be a length-2 vector representing a single location at which features will be queried, or a list of two coordinates representing the bottom-left and top-right corners of a bounding box within which features will be queried. Defaults to the current map view.}

\item{layers}{A vector of layer names to include in the query}

\item{filter}{A filter expression used to filter features in the query.}
}
\value{
A list of features in GeoJSON format, accessible at \code{input$MAPID_feature_query} in your Shiny app code.
}
\description{
Query rendered features on a map in a Shiny session
}
