% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{renderCompare}
\alias{renderCompare}
\title{Render a comparison map in a Shiny application}
\usage{
renderCompare(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a comparison map.}

\item{env}{The environment in which to evaluate the expression.}

\item{quoted}{Logical, whether the expression is quoted. This is useful if you want to save an expression in a variable.}
}
\value{
A rendered comparison map.
}
\description{
This function renders a comparison map in a Shiny application.
}
