% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{set_filter}
\alias{set_filter}
\title{Add a filter to a map using a proxy}
\usage{
set_filter(proxy, layer_id, filter)
}
\arguments{
\item{proxy}{A proxy object created by \code{mapboxgl_proxy} or \code{maplibre_proxy}.}

\item{layer_id}{The ID of the layer to which the filter will be applied.}

\item{filter}{The filter expression to apply.}
}
\value{
The updated proxy object.
}
\description{
This function allows a filter to be added to an existing Mapbox GL or Maplibre GL map using a proxy object.
}
