% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{set_filter}
\alias{set_filter}
\title{Set a filter on a map layer}
\usage{
set_filter(map, layer_id, filter)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function, or a proxy object.}

\item{layer_id}{The ID of the layer to which the filter will be applied.}

\item{filter}{The filter expression to apply.}
}
\value{
The updated map object.
}
\description{
This function sets a filter on a map layer, working with both regular map objects and proxy objects.
}
