% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{set_source}
\alias{set_source}
\title{Set source of a map layer}
\usage{
set_source(map, layer_id, source, ...)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function, or a proxy object.}

\item{layer_id}{The ID of the layer to update.}

\item{source}{An sf object (which will be converted to a GeoJSON source).}

\item{...}{Additional arguments for compatibility with versions <= v0.2.1.}
}
\value{
The updated map object.
}
\description{
Set source of a map layer
}
