% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{set_source}
\alias{set_source}
\title{Set source of a map layer}
\usage{
set_source(map, layer, source)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function, or a proxy object.}

\item{layer}{The ID of the layer to update.}

\item{source}{The source as geojson}
}
\value{
The updated map object.
}
\description{
Set source of a map layer
}
