% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{set_tooltip}
\alias{set_tooltip}
\title{Set a paint property on a map layer}
\usage{
set_tooltip(map, layer, tooltip)
}
\arguments{
\item{map}{A map object created by the \code{mapboxgl} or \code{maplibre} function, or a proxy object.}

\item{layer}{The ID of the layer to update.}

\item{tooltip}{The name of the new tooltip variable which should be set.}
}
\value{
The updated map object.
}
\description{
Set a paint property on a map layer
}
