% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{turf_convex_hull}
\alias{turf_convex_hull}
\title{Create convex hull}
\usage{
turf_convex_hull(
  proxy,
  layer_id,
  result_layer_id = NULL,
  layer_style = NULL,
  send_to_r = FALSE
)
}
\arguments{
\item{proxy}{A mapboxgl_proxy or maplibre_proxy object.}

\item{layer_id}{The ID of the layer containing the points.}

\item{result_layer_id}{Optional. The ID for the new layer to display the convex hull on the map.}

\item{layer_style}{Optional. A named list of paint properties for styling the result layer.}

\item{send_to_r}{Logical. If TRUE, sends the result back to R as input$<map_id>_turf_result.}
}
\value{
The proxy object for method chaining.
}
\description{
This function creates a convex hull around a set of points.
}
