% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{turf_voronoi}
\alias{turf_voronoi}
\title{Create Voronoi diagram}
\usage{
turf_voronoi(
  proxy,
  layer_id,
  bbox = NULL,
  result_layer_id = NULL,
  layer_style = NULL,
  send_to_r = FALSE
)
}
\arguments{
\item{proxy}{A mapboxgl_proxy or maplibre_proxy object.}

\item{layer_id}{The ID of the layer containing the points.}

\item{bbox}{Optional. A vector of four numbers representing the bounding box \link{minX, minY, maxX, maxY}.}

\item{result_layer_id}{Optional. The ID for the new layer to display the Voronoi diagram on the map.}

\item{layer_style}{Optional. A named list of paint properties for styling the result layer.}

\item{send_to_r}{Logical. If TRUE, sends the result back to R as input$<map_id>_turf_result.}
}
\value{
The proxy object for method chaining.
}
\description{
This function creates a Voronoi diagram from a set of points.
}
